Implement product management and exception handling
Added serialization support to model classes and DAOs, implemented exception classes, and extended the controller to handle product registration, user authentication, and category management. Refactored DAOs to improve error handling during file operations and updated model fields for consistency. Added new .secn files for product categories and users.

+++ proyecto/src/co/edu/unbosque/view/CompraProducto.java
@@ -17,7 +17,7 @@ public class CompraProducto extends JPanel {
     
     private JButton info;
 
-    public CompraProducto(String nombre, float precio, String rutaFoto, int id, ActionListener e) {
+    public CompraProducto(String nombre, double precio, String rutaFoto, int id, ActionListener e) {
 
     	this.setPreferredSize(new Dimension(250, 370));
         this.setBackground(Color.WHITE);
